Refactor modules and add local Postgres compose
