+++ compose.yaml
@@ -1,0 +1,21 @@
+services:
+  postgres:
+    image: postgres:16-alpine
+    container_name: mr-roller-postgres
+    restart: unless-stopped
+    environment:
+      POSTGRES_DB: mr_roller
+      POSTGRES_USER: mr_roller
+      POSTGRES_PASSWORD: secret
+    ports:
+      - "5432:5432"
+    volumes:
+      - mr_roller_postgres_data:/var/lib/postgresql/data
+    healthcheck:
+      test: ["CMD-SHELL", "pg_isready -U mr_roller -d mr_roller"]
+      interval: 5s
+      timeout: 5s
+      retries: 10
+
+volumes:
+  mr_roller_postgres_data:
