+++ mr-roller-discord/src/render.rs
